feature: avoid debug logging private keys
