+++ bip32/src/lib.rs
@@ -77,7 +77,10 @@ extern crate lazy_static;
 
 #[cfg_attr(tarpaulin, skip)]
 #[macro_use]
-pub(crate) mod prelude;
+pub(crate) mod macros;
+
+/// Everything needed for common usage
+pub mod prelude;
 
 /// Low-level types
 pub mod primitives;
